changed datasetsDto to used list double instead of list int

+++ DSUGrupp1/Models/DTO/DatasetsDto.cs
@@ -5,12 +5,12 @@
         public DatasetsDto() 
         {
             Label = "test";
-            Data = new List<int> { 1, 2, 3 };
+            Data = new List<double> { 1, 2, 3 };
             BackgroundColor = new List<string> { "rgb(220, 174, 198)" };
             BorderWidth = 10;   
         }
         public string Label { get; set; }
-        public List<int> Data { get; set; }
+        public List<double> Data { get; set; }
         public List<string> BackgroundColor { get; set;}
         public int BorderWidth { get; set;}
